pyproject.toml
[tool.poetry]
name = "messagessendernew"
version = "0.1.0"
description = ""
authors = ["CatritMe"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-apscheduler = "^0.6.2"
django = "4.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
